refactor(api): streamline imports and enhance Swagger documentation
- Updated import paths for the list router to improve code organization.
- Enhanced Swagger documentation by adding detailed tags and tag groups for better API categorization.
- Standardized tag names for authentication and POI management to improve clarity and consistency across the API.
